requirements.txt
discord.py==2.6.4
dotenv==0.9.9
ipykernel==7.1.0
jupyter==1.1.1
matplotlib==3.10.7
notebook==7.4.7
numpy==2.3.4
pandas==2.3.3
editdistance==0.5.3